requirements.txt
asgiref==3.4.1
boto3==1.20.11
botocore==1.23.11
dj-database-url==0.5.0
Django==3.2.9
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-js-asset==1.2.2
django-on-heroku==1.1.2
django-storages==1.12.3
gunicorn==20.1.0
jmespath==0.10.0
Pillow==8.4.0
psycopg2-binary==2.9.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
